requirements.txt
fastapi==0.115.0
uvicorn==0.30.6 
passlib[bcrypt]==1.7.4
pydantic==2.7.3
# python-jose[cryptography]==3.3.0 
python-dotenv==1.0.0
jinja2==3.1.4
bcrypt==3.2.0
pymongo==4.5.0
motor==3.3.1




#venv\Scripts\activate

#uvicorn main:app --reload